Additions to graph, fixed bug in SwingOps (view could be garbage collected)
Added start of box selection to graphs, added more refs passed to
GraphBasic to allow for external control of mode, zooming etc. BoxesDemo
graph usage now has code for a relatively full graph with toggle buttons
to select either zoom mode or select mode, a button to zoom out (auto
zoom), etc. Fixed annoying bug where SwingOps would sometimes stop
responding - was due to not retaining a reference to the View that
enables/disables the action, so that this view could be garbage
collected, leaving enabled state frozen. This is an undesirable but
inevitable product of not keeping strong links to Views, to avoid memory
leaks. Essentially, you must remember to keep a reference from anything
that cares about a View continuing to operate, to that View. Also, now
ignoring workspace.xml, hopefully this will stick.