Better implementation of background reaction
Background reactions are now based on a similar pattern to Reactions - a
responseSource is used, which performs whatever Box reads are required,
and then returns a function that will carry out the background
processing and (optional) writes to Boxes. This means that the data
gathered by the Box reads doesn't need to be messily externally exposed
as a "data" object, and makes most code neater. It also makes Background
reactions look more like normal Reactions, and will allow for an adapter
to run Reactions as (non-cancelling) background reactions.