Added demo for editing an element of a list. Fixed issue with StringOptionView when underlying Var changes.
The issue with the StringOptionView shows up when the Swing component
loses focus as a result of (for example) a click that selects a new Var
to be edited. When this happens, the component commits in the swing
thread AFTER the Var has changed, and then overwrites the change. In an
extreme case it may even be writing to a PathVar, and hence change the
wrong data. This can be avoided by making sure that the view doesn't
commit unless it still sees the same data it was displaying before the
editing started.