Implemented a much better solution to StringOptionView editing issue.
Uses replacing/adding of Swing updates for committing and displaying
data so that displaying always takes precedence. This allows for a
chance for the View clause to make sure the StringOptionView will just
display new data, rather than trying to commit over the top of it. This
replaces the previously committed partial solution.