Fixed some related list/path/view bugs. Made NumberOptionView provide an implicit sequence for each numeric type.
The bugs were related to Path handling: 1 - Path reactions were intended
to copy endpoint of path to the path var when the endpoint had more
recent data, and from path var to endpoint of path if path var had more
recent data (defaulting to copying endpoint to path var if not certain).
However the path reaction itself can hence modify the path var, making
it look like it has the more recent data (since the endpoint var may not
ever have a change - it may be a change to the path itself that causes
the reaction to be called). To avoid overwriting valid endpoint data in
this case, the path reaction can now ONLY copy data from path var to
endpoint var the first time it is called each cycle. This means it is
never basing the direction of propagation on changes it has made itself.
It also means that data is only copied from path var to endpoint var if
the path var is changed before the endpoint var AND before anything in
the path. This is a much safer strategy, and in the case where there are
no reactions targeting the path var, it is always correct. Path vars are
extremely useful, but still present some problems - it might be best to
remove their local storage and make reads and writes pass directly
through to the endpoint, if this is possible. 2 - This bug only showed
up when using paths, but might occur otherwise - when the user clicks
out of a Swing View that is displaying something like a Path var, where
the view has stored state, and into a table, the table receives the
click and hence may change state BEFORE the swing view loses focus. The
swing view then commits itself into a NEW var, which is probably not
valid. To prevent this, the view should lose focus BEFORE the table
gains it and is edited. Unfortunately, swing doesn't seem to sequence
things like this. To work round this, out list selection models now use
SwingView.addUpdate to make changes to the selection Var, so that the
changes are delayed until after other swing components (including the
problematic Swing Views) have a chance to react appropriately.