More robust path implementation
Reviewed the path problem, and implemented a much simpler solution that
should also be more robust - as far as I can tell at the moment, this is
now completely "correct".  Following the conventions in the code, there
is a Var v which "mirrors" another Var e. These should always contain
the same value, and writes to one result in the same value being written
to the other. The complication is that Var e is dynamic - it is located
by executing a path function that returns the current Var that we are
treating as "e". The path function will almost always read one or more
boxes when it evaluates. Previously, we ignored these boxes, but now we
are aware of changes to them, when deciding whether to copy the value of
e to v, or vice versa, in the Path Reaction. Essentially, we only copy
from v to e iff v has changed, but neither e nor any of the boxes used
by the path function has changed. So writes directly to v, or via
reactions that do NOT change e or the path function boxes, will be
copied to e. Otherwise, e will be copied to v. So as long as the path is
stable, and e remains the same, v is free to write to it. Whenever the
path changes, v defaults to just mirroring the new values of e.