New Reaction trait
Reaction now has a method react() that can do reading and writing of
state. This replaces the old respond() that only did reading, but
returned a function that would do writing. This separation was thought
to be necessary earlier in implementation, to allow "cancellation" of
reactions by not applying the function, but this is no longer needed.
It is much easier to write mixed methods that read and write together.

To convert existing reactions, they have been given a react()
implementation that is just respond().apply(). This can be removed over
time, and the old respond() methods rewritten to just perform their
writes rather than returning functions. However this needs to be
thorough, since the compiler will NOT give an error if any of the
functions are missed - this will just create a pointless function then
lose it, rather than performing the action. Note that there are
anonymous reactions as well, making this harder to find, since they do
not have return types specified.

Once all reactions are converted, some methods accepting : => (() =>
Unit) can be converted to just : => Unit to make things easier.

Note, have also removed applyingReaction var from Box, and made
beforeWrite return Unit, since there is no current use of the boolean
value it used to return.