New Ledger system, and reflection fix in Node.
Node.accessorMap used declared methods of the node class, should use
all methods.

The old system for Ledgers made Ledgers themselves Boxes, but very
non-standard boxes in that they had no single value for contents. In
addition, the RecordView had to be an odd Box.

The new system uses immutable Ledgers stored in a Var, in much the same
way that we use immutable Lists in a ListVar. This requires some
changes to code using Ledgers, for example a Var(ListLedger(...))
becomes a ListLedgerVar(...)

Things like ListLedger are now split in two - firstly the new immutable
ListLedger handles data indexing, but not change logging. Then we have
a ListLedgerVar object that creates a plain LedgerVar with a reaction
that updates the ListLedger in the LedgerVar so that it is always
constructed from the contents of two source Vars, one holding a list,
and one holding a new immutable RecordView.

Overall, this makes the system more consistent between lists and
ledgers, gives more immutable data without reducing capabilities, and
allows for making all Boxes into Refs, in the sense that all Boxes can
contain a single value - this could be very useful for e.g. an
undo-redo system, which could rely on all Changes being SingleChanges,
and all changes being undoable by just setting the old value into the
Box/Ref.

Testing with demos, etc. indicates that the performance of the new
system is comparable to the old one, and usage is very similar.

Added maven dependency on scalaz, for experimentation.

Added map and flatMap for ListRef.