Unified changes
All Boxes now need C to be a Change, where Change gives an old value
and a new value.
Modified Ledger to have such changes
SingleChange now extends Change
Various parametric types have to be constrained to be C <: Change[T]
Changed some asInstanceOf's to type ascription.
In general it is now just a pain to use Box[T, _] instead of Ref[T],
and likewise Box[List[T], _] instead of ListRef[T], and similarly for
VarBox and Var. The benefit was never really that obvious - only Ref
really makes sense in context, the extra generality of Box would never
be used, since the parameter already threw away the change type. Note
however that this does mean that users will need to use ListRef,
ListCal etc. wherever they can, which is a good idea anyway, in case in
the future we want to use the additional change information in a
ListRef to optimise calculations etc.
ListIndices have adapted to ignore the extra oldValue and newValue
parts of changes.ListChange trait now extends Change[List[T]]
Removed ListRef.map, will be restored when rewritten to adapt
ListChanges (we now need to transform the old and new values
appropriately to generate changes)