Rough start to JFX graph support
Boxes graphing was designed to work across multiple graphical backends,
have started implementation of a "direct mode"wrapper to use JavaFX
nodes via a GraphCanvas interface. This converts drawing operations to
nodes, and adds them to a "buffer"ndoe which can then replace the
graph's children. This doesn't work yet, but checking in as a basis for
further work!