Added start of JSON persistence
Simple DataTarget, prints more or less JSONish output, with slightly
odd handling of tags as names, and class tags as whole objects, with
type, id and ref as names as well. This means that when a tag is opened
and immediately closed, it needs a dummy value, currently 1.
Parsing may be trickier. Ultimately this should lead to BSON data,
which can be stored in mongodb, etc.