New InstanceViewSource behaviour
Altered to use a new WeakRef to handle the problem of not retaining
viewed objects. This slightly violates Box contract, but in a way that
is valid for the limited use case we have here. Still assessing whether
this is a better way to do things - it has a big advantage of causing
less changes when the model of the MultiView changes, and also avoids
resetting Views that are not being used to default models, which causes
loss of e.g. list selections, etc.